# Increase perf for parsing remote maps