This reduces the wait time by 33-37%

Bleed:
21393 ms RemoteMapDetails
13891 ms RemoteMapDetails

PR:
13479 ms RemoteMapDetails
6852 ms RemoteMapDetails