@pchote 2025-03-30
Note that the perf improvement will be even bigger than the numbers suggest because it will update maps as their info arrives instead of waiting until the end.